VirtioNet: Changed HashMap implementation to arrays
Due to Rust's rules, we would lose ownership of our packets when we give
them to the device.

Previously, we prevented this by moving them into a HashMap with the buffer
address as the key and the buffer as the value.

However, this isn't performant. Taking inspiration from `ixgbe`, we used
an array of `Option<RRef<NetworkPacketBuffer>>`. We index into this
array by what descriptor it's allocated as.